# Use optimized path of setting dataset size during output discovery